list_upgrades: download packages and populate dom0 metadata
Dom0 doesn't have direct access to (online) repository metadata and
since most of Fedora repositories don't have metadata signed, we don't
want to expose dom0 to it. Instead, to check updates, download all the
updates in UpdateVM, transfer them to dom0 (where signature on packages
will be checked) and populate dom0 local repository metadata. Then
normal queries (like `dnf check-update`) will work.

The drawback is that it downloads all the packages just for listing
them, but in most common case, those packages will be installed in a
moment anyway and will not be downloaded for the second time for it.

Fixes QubesOS/qubes-issues#2089