Allow buster changing suite from testing (to stable)
Apt in buster refuse to accept Release file with a different suite name
than it had before. Generally it is a good thing, but unfortunately it
breaks update of buster installed when it was still testing.
Detect this by looking for 'Suite: testing' in apt cached list files,
and if found, pass --allow-releaseinfo-change to apt-get update command.

Ideally, we'd like to allow explicitly testing->stable change for
buster only, but there is no such option in apt-get (yet?).

Fixes QubesOS/qubes-issues#5149